+++ src/main/java/com/ticketscale/infrastructure/security/SecurityConfigurations.java
@@ -30,6 +30,8 @@ public class SecurityConfigurations {
                 .authorizeHttpRequests(req -> {
                     req.requestMatchers(HttpMethod.POST, "/login").permitAll();
                     req.requestMatchers(HttpMethod.POST, "/usuarios").permitAll();
+                    req.requestMatchers(HttpMethod.POST, "/eventos").hasRole("ADMIN");
+                    req.requestMatchers(HttpMethod.DELETE, "/eventos/**").hasRole("ADMIN");
                     req.anyRequest().authenticated();
                 })
                 .addFilterBefore(securityFilter, UsernamePasswordAuthenticationFilter.class)